Return decimal type as string
If you have a decimal with multiple trailing zeros, e.g. 2 for monetary fields or more, they are currently truncated to 1 precision point due to the float conversion. Since the encoding returns a string type anyway, and the conditional statement checks for decimal type, we can safely convert to string type here.