Do not try to deserialize all strings to JSON
Previously, the JSONField would try to interpret all strings handed into it as JSON, and pass it to json.loads to deserialize the.

This fails, however, if we want to pass a simple String value in to the field, because that gets interpreted as JSON, and then the JSON serialization fails.

As a remedy, this change updates the JSONField to only interpret strings beginning with `"`, `[`, and `{` as JSON, and save all other strings directly. This still has some edge cases (what if I want to save a this string: `"[hello]"`?) but works for the majority of use cases.